% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_package_ChooseRef.R
\name{Diss_Ref_Plot}
\alias{Diss_Ref_Plot}
\title{Title}
\usage{
Diss_Ref_Plot(
  RELEVES,
  REF,
  DISTANCES,
  METHOD = "euclidean",
  COUL_RELEVES = 2,
  COUL_Rel_variable = TRUE,
  COUL_REF = 1,
  COUL_Seg = "#9C8809",
  COUL_Seg_variable = TRUE,
  CHOIX_NOMBRE = "absent",
  N_REF = "absent",
  DIST_MIN = "absent",
  VAL_DIST = TRUE,
  DECAL = 0
)
}
\arguments{
\item{RELEVES}{-}

\item{REF}{-}

\item{DISTANCES}{r}

\item{METHOD}{r}

\item{COUL_RELEVES}{r}

\item{COUL_Rel_variable}{r}

\item{COUL_REF}{r}

\item{COUL_Seg}{r}

\item{COUL_Seg_variable}{r}

\item{CHOIX_NOMBRE}{r}

\item{N_REF}{r}

\item{DIST_MIN}{r}

\item{VAL_DIST}{r}

\item{DECAL}{r}
}
\description{
Calcul de l'erreur standard
}
