% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_package_ChooseRef.R
\name{DissRef3}
\alias{DissRef3}
\title{Similarities between references and restoration}
\usage{
DissRef3(RELEVES, REF, METHOD = "bray", BINARY = FALSE, DOUBLONS = TRUE)
}
\arguments{
\item{RELEVES}{Variables of the restoration sites data matrix}

\item{REF}{Variables surveys of reference sites data matrix}

\item{METHOD}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao" or "mahalanobis". Cf \code{\link{vegan::vegdist}} function}

\item{BINARY}{(FALSE by default) If TRUE, data is converted to binary data}

\item{DOUBLONS}{(TRUE by default) Are REF plots also included in RELEVES data?}
}
\value{
\item{Diss_Mean}{average dissimilarity by RELEVES}

\item{Diss_Min}{minimal dissimilarity by RELEVES}

\item{RelRef_order}{list of REF names in ascending order of dissimilarity for each RELEVES}

\item{DistRef_order}{list of REF distance values in ascending order of dissimilarity for each RELEVES}
}
\description{
Calculate similarity between plots and a group of reference plot
}
