% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_package_ChooseRef.R
\name{Diss_Ref_Plot}
\alias{Diss_Ref_Plot}
\title{Plot of references and restoration site dissimilarities}
\usage{
Diss_Ref_Plot(
  RELEVES,
  REF,
  DISTANCES,
  METHOD = "euclidean",
  COUL_RELEVES = 2,
  COUL_Rel_variable = TRUE,
  COUL_REF = 1,
  COUL_Seg = "#9C8809",
  COUL_Seg_variable = TRUE,
  CHOIX_NOMBRE = "absent",
  N_REF = "absent",
  DIST_MIN = "absent",
  VAL_DIST = TRUE,
  DECAL = 0
)
}
\arguments{
\item{RELEVES}{Variables of the restoration sites data matrix}

\item{REF}{Variables of reference sites data matrix}

\item{DISTANCES}{Object resulting from the analysis of the \code{\link{ChosseRef::DissRef3}} function.}

\item{METHOD}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao" or "mahalanobis". Cf \code{\link{vegan::vegdist}} function}

\item{COUL_RELEVES}{Color of RELEVES names}

\item{COUL_Rel_variable}{If TRUE: color change for each RELEVES}

\item{COUL_REF}{Color of REF names}

\item{COUL_Seg}{Color of links}

\item{COUL_Seg_variable}{If TRUE: color change for each link}

\item{CHOIX_NOMBRE}{Display of the reference number \code{N_REF} or display of the distance value \code{DIST_MIN}.}

\item{N_REF}{Number of REF to display a link between REF and RELEVES}

\item{DIST_MIN}{Minimum distance value to display a link between REF and RELEVES}

\item{VAL_DIST}{Distance value display}

\item{DECAL}{Distance between the printing of the distance and the link}
}
\description{
Display similarity link (previously calculated with \code{\link{ChosseRef::DissRef3}} function) between plots and a group of reference plot
}
